Move downloader to downloader.js
May be easier to locate with IDE path fuzzy search
